HBASE-24132 Upgrade to Apache ZooKeeper 3.5
In CDPD-3352 we upgraded to ZooKeeper 3.5.5, but that time the upstream HBase community was still
committed to use the old 3.4 ZooKeeper so we had to make downstream-only change. Later the upstream
HBase upgraded to ZooKeeper 3.5.7 on branches: master, branch-2, branch-2.3.

With backporting this change we are fixing CDPD-15203, avoiding IllegalStateException in ZKWatcher
("Received event is not valid: Closed").

Important note just for clarification: we haven't upgraded the donwstream ZooKeeper version
for HBase to 3.5.7 with this commit, but we still use the CDP ZooKeeper 3.5.5 in HBase.

Co-authored-by: 姜建飞 10222269 <jiang.jianfei@zte.com.cn>

Signed-off-by: Mate Szalay-Beko <symat@apache.org>
Signed-off-by: Norbert Kalmar <nkalmar@apache.org>
Signed-off-by: stack <stack@apache.org>
Signed-off-by: Duo Zhang <zhangdo@apache.org>

Change-Id: Iff9322815e70f10efe3108c80708b40d47c425e8
Author: Jianfei Jiang
Reason: Improvement
Ref: CDPD-15203
(cherry picked from commit a029b6ffb14e46bbe33a71d84d25d385dba0fc5b)